+++ cobs/util/parallel_for.hpp
@@ -22,10 +22,10 @@ extern std::unique_ptr<tlx::ThreadPool> g_thread_pool;
 
 //! run a functor in parallel
 template <typename Functor>
-void parallel_for(size_t begin, size_t end, size_t num_threads,
+void parallel_for(uint64_t begin, uint64_t end, uint64_t num_threads,
                   Functor functor) {
     if (num_threads <= 1) {
-        for (size_t i = begin; i < end; ++i) {
+        for (uint64_t i = begin; i < end; ++i) {
             functor(i);
         }
     }
@@ -34,14 +34,14 @@ void parallel_for(size_t begin, size_t end, size_t num_threads,
             g_thread_pool = std::make_unique<tlx::ThreadPool>();
 
         tlx::Semaphore sem;
-        std::atomic<size_t> counter { begin };
+        std::atomic<uint64_t> counter { begin };
         std::exception_ptr eptr;
         // enqueue threads for work
-        for (size_t t = 0; t < num_threads; ++t) {
+        for (uint64_t t = 0; t < num_threads; ++t) {
             g_thread_pool->enqueue(
                 [&]() {
                     try {
-                        size_t i;
+                        uint64_t i;
                         while ((i = counter++) < end) {
                             functor(i);
                         }
